+++ CHANGES.rst
@@ -17,6 +17,7 @@ Features
 
 - Added option `--rerun-except` to rerun failed tests those are other than the mentioned Error.
 
+- Expose `reruns` and `reruns_delay` through `pytest.ini` file.
 
 10.2 (2021-09-17)
 -----------------